+++ libs/components/src/index.ts
@@ -13,15 +13,23 @@ export * from './lib/card/card.component';
 export * from './lib/modal/modal.component';
 export * from './lib/tabs/tabs.component';
 export * from './lib/tabs/tab.component';
+export * from './lib/accordion/accordion.component';
+export * from './lib/accordion/accordion-item.component';
+export * from './lib/divider/divider.component';
 
 // Data Display Components
 export * from './lib/badge/badge.component';
 export * from './lib/avatar/avatar.component';
 export * from './lib/tooltip/tooltip.component';
+export * from './lib/chip/chip.component';
 
 // Feedback Components
 export * from './lib/alert/alert.component';
 export * from './lib/spinner/spinner.component';
 export * from './lib/progress/progress.component';
+export * from './lib/skeleton/skeleton.component';
+
+// Navigation Components
+export * from './lib/breadcrumb/breadcrumb.component';
 
 // More components will be exported as they are implemented